==> PANDAS/1-Series_y_DataFrames.ipynb <==
import numpy as np
import pandas as pd

Ejercicio 1. ¿Qué diferencia hay entre una Serie y un DataFrame?

# Las Series tienen una dimensión(índices) y los DataFrames tienes 2 dimensiones (índices y columnas).

Ejercicio 2. Momento de trabajar con Series.

valores = list(np.random.randint(1,50,(7)))
valores

indices = ["Lunes", "Martes", "Miercoles", "Jueves", "Viernes", "Sabado", "Domingo"]
indices

# convertimos las dos listas a serie

semana = pd.Series(valores, indices)
semana

# A partir de las dos listas anteriores cread un diccionario usando un bucle for. 
# Hemos creado el diccionario de dos formas. Usaremos para el ejercicio la última "diccionario = dict(zip(indices, valores))" porque es la que nos
# ha resultado más comoda para trabajar, pero a continuación dejamos otro método con el bucle for y una list comprehension.

keys1 = indices
values1 = valores


lista_del_zip = list(zip(keys1, values1))
lista_del_zip

dicc2 = {k: v for k, v in lista_del_zip}

dicc2

# Continuámos con el ejercicio.
# A partir de las dos listas anteriores cread un diccionario usando un bucle for. 
# Para hacer esto tenemos el método zip que nos permite iterar por dos listas a la vez.

diccionario = dict(zip(indices, valores))
diccionario

# Crea una Serie con el diccionario creado en el punto anterior.

semana2 = pd.Series(diccionario)
semana2

# ¿Qué forma tiene esta última Serie?

semana2.shape

# ¿Qué tipo de datos tiene?

semana2.dtypes

# ¿Qué tamaño tiene?

semana2.size

# ¿Cuáles son los índices de la Serie?

semana2.index

# ¿Y los values?

semana2.values

# Extraed el valor para el índice "lunes".

semana2["Lunes"]

# Extraed el valor para el índice "lunes", "martes" y "miércoles".

semana2["Lunes" : "Miercoles"]

# Extraed el valor para el índice "lunes" y "domingo".

semana2[["Lunes", "Domingo"]]

Ejercicio 3. Es el turno de los DataFrames

# Cread un array de dos dimensiones con 4 filas y 3 columnas.

array1 = np.random.randint(1,50,(4, 3))
array1

# Convertid el array en un DataFrame donde las columnas se llamen "España", "Francia" e "Italia".

paises = pd.DataFrame(array1, columns =["España", "Francia", "Italia"])
paises

# Descargad el archivo "medallas" y cargad el DataFrame con la siguiente línea de código:

df = pd.read_csv("medallas.csv", index_col=["País"])


# Usando el loc extraed todas las columnas de la fila de Tailandia.

df.loc["Tailandia"]

# Usando el loc extraed todas las columnas de China e Irán.

df.loc[["China", "Irán"],]



# Usando el loc extraed solo las columnas de "Oro", "Plata" y "Bronce" de China e Irán.

df.loc[["China", "Irán"], ["Oro", "Plata", "Bronce"]]

df

# Usando el iloc extraed la información de la India. Devolved todas las columnas.

df.iloc[5,:]

# Usando el iloc extraed todas las columnas de los paises que estén en posición impar. La primera que nos tendría que salir es Japón y la última Corea del Norte.

df.iloc[1::2, :]

# Utilizando el código del ejercicio anterior, seleccionad unicamente las columnas "Oro", "Plata" y "Bronce".

df.iloc[1::2, 1:]

# Contestad a las siguientes preguntas:
# ¿Qué país consiguió más de 200 medallas de oro?

df.loc[(df.Oro > 200), 'Oro']

# ¿Qué país consiguió más de 200 medallas de oro y más de 1000 de plata?

df.loc[(df.Oro > 200) & (df.Plata > 1000), ['Oro', 'Plata']]

# ¿Qué país consiguió más de 200 medallas de oro o más de 700 de bronce?

df.loc[(df.Oro > 200) | (df.Bronce > 700), ['Oro', 'Bronce']]

# Como vemos hay una columna que se llama Unnamed: 0 que no nos interesa. Eliminad esa columna

del df["Unnamed: 0"]

df

# Cread una nueva columma que se llame "total" que sea el resultado de la suma de todas las medallas obtenidas por cada país.

df_total = ["Oro", "Plata","Bronce"]

df.insert(3, "Total", df[df_total].sum(axis=1))

df
