==> PANDAS/2-Carga_y_guardado_de_datos.ipynb <==
Id cargando todos los ficheros en vuestro jupyter.

import pandas as pd
import os 
import numpy as np

archivo_csv = pd.read_csv("DATOS/quejas.csv", encoding = "latin-1", sep = ";")
archivo_csv

archivo_json = pd.read_json("DATOS/quejas.json")
archivo_json

archivo_xlsx = pd.read_excel("DATOS/quejas2.xlsx")
archivo_xlsx

archivo_pickle = pd.read_pickle("DATOS/algas.pickle")
archivo_pickle

archivo_pepinillo = pd.read_clipboard(sep = "\t")
archivo_pepinillo

archivo_pepinillo.to_json("DATOS/pepinillos.json")
archivo_pepinillo.to_excel("DATOS/pepinillos.xlsx")
archivo_pepinillo.to_pickle("DATOS/pepinillos.pickle")
archivo_pepinillo.to_csv("DATOS/pepinillos.csv")

archivo_pepinillo_2 = pd.read_clipboard(sep = "\t")
archivo_pepinillo_2

archivo_panda = pd.read_clipboard(sep = ":")
archivo_panda

archivo_panda.reset_index(drop=True)

archivo_panda.to_pickle("DATOS/panda.pickle")

archivo_panda_cambiado = np.swapaxes(archivo_panda,1,0)
archivo_panda_cambiado

archivo_panda_cambiado.shape
