==> PANDAS/3-Métodos_de_pandas.ipynb <==
Cargar el dataset

import pandas as pd

fichero = pd.read_csv("DATOS/attacks.csv", sep = "," )

Haced una exploración básica del DataFrame:

¿Cuántas filas y columnas tiene el DataFrame?

fichero_shape = fichero.shape
fichero_shape

Mostrad las primeras cinco filas y las últimas cinco filas del DataFrame.

display(fichero.head(), fichero.tail())

Mostrad 2 filas al azar del DataFrame

fichero.sample(2)

¿Cuál es el nombre de las columnas del DataFrame?

fichero.columns

¿Cuáles son los principales estádisticos del DataFrame? Mostrad por separado las variables categóricas y las variables numéricas.

#estadísticas numéricas
fichero.describe()

#estadísticas categóricas 
fichero.describe(include="object")

¿Cuál es el porcentaje de valores nulos en cada columna?

fichero_porcentaje = round((fichero.isnull().sum() /fichero.shape[0]) * 100 , 2)
fichero_porcentaje

Sacad los valores únicos y su frecuencia de las variables numéricas. Almacenad los resultados en un diccionario. Uno para cada variable.

#Primero miramos cuales son las columnas de tipo númerico
fichero.dtypes

fichero_unname = fichero["Unnamed: 0"].value_counts()
fichero_unname

diccionario_unname = dict(zip(fichero_unname.index, fichero_unname.values))
diccionario_unname

fichero_year = fichero["Year"].value_counts()
fichero_year

diccionario_año = dict(zip(fichero_year.index, fichero_year.values))
diccionario_año

fichero_order = fichero["original order"].value_counts()
fichero_order

diccionario_order = dict(zip(fichero_order.index, fichero_order.values))
diccionario_order

Cambiad el nombre de las columnas para que todas estén en minúsculas y reemplazad los espacios por _.

nuevas_columnas = {columna : columna.lower().replace(" ", "_") for columna in fichero.columns }
nuevas_columnas

fichero.rename(columns=nuevas_columnas, inplace= True)
fichero.columns

Eliminad las columnas de Unnamed: 22 y Unnamed: 23, Case Number.1, Case Number.2, original order, Investigator or Source, pdf, href formula.

fichero_sin_columnas = fichero.drop(["unnamed:_22","unnamed:_23", "case_number.1", "case_number.2", "original_order", "investigator_or_source", "pdf", "href_formula" ], axis=1)
fichero_sin_columnas.head(1)

Establece la columna Case Number como índice del DataFrame.

fichero_sin_columnas.set_index(['case_number'], inplace= True)

fichero_sin_columnas.head()

Guarda el csv que no tiene esas columnas para seguir usándolo mañana.

fichero_sin_columnas.to_csv("DATOS/attacks_sin_columnas.csv")
