==> PANDAS/4-Filtrado_de_datos.ipynb <==
import pandas as pd
import numpy as np

1.  Cargad el csv que guardasteis ayer. Antes de seguir con las preguntas, este DataFrame está terriblemente sucio. En las próximas sesiones lo iremos limpiando. 

Para evitar posibles errores con los nulos vamos a eliminar todos los nulos del DataFrame.Para ello vamos a utilizar el método .dropna() de pandas. Tendréis que ejecutar la siguiente línea de código para eliminar los nulos del DataFrame y poder seguir haciendo el resto de los ejercicios.

df.dropna(inplace = True)

    ¿Cuántas filas se han eliminado con esta línea de código?

fichero = pd.read_csv("DATOS/attacks_sin_columnas.csv")

fichero.info()
#en el fichero hay 6114 entry.

fichero.dropna(inplace = True)

fichero.info()

#Hemos eliminado 4612 entradas.

2. ¿Cuál es el número de ataques que se han producido en Estados Unidos?

fichero.head(2)

#hemos llamado al dataframe para averiguar donde encontramos la informacion correspondiente a USA, si en country, location o area.

estados_unidos = fichero[fichero["country"] == "USA"]
estados_unidos.head(2)

#Tenemos 719 registros de USA, ponemos un .head para que se vea mejor.

3. ¿Y en España?
        

fichero["country"].unique()

# Hacemos un unique para averiguar como llaman a España

españa = fichero[fichero["country"] == "SPAIN"]
españa.head(2)

#En España tenemos 5 ataques

-   ¿En qué regiones se han producido los ataques (columna location) ?
-   ¿Cuántos ataques se han producido en cada área (columna area?

localidad = españa["location"].unique()
localidad

localidad_conteo = españa["location"].value_counts()
localidad_conteo

Extraed un DataFrame nuevo con los ataques que se han producido en Australia y Estados Unidos.

4.1 ¿Cuántos ataques se han producido en cada país?


        
    📌 Nota Este ejercicio se puede resolver usando dos condiciones de filtrado o usando el método isin(). Hacedlo con ambas herramientas.

australia_usa = fichero[(fichero["country"] == "USA") | (fichero["country"] == "AUSTRALIA")]
australia_usa.head(2)

#tenemos 1043 registros

australia_usa_isin = fichero[fichero["country"].isin(["AUSTRALIA","USA"])]
australia_usa_isin.head(2)

#tenemos 1043 registros, entonces el mismo resultado de antes.

australia = australia_usa[australia_usa["country"] == "AUSTRALIA"]
australia.head(2)

#Se han producido 324 ataques en Australia

estados_unidos.head(2)

#Ya teniamos la variable creada y aprovechamos para averiguar el numero de ataque que son 719.

- 4.2 ¿Ha habido ataques en Australia a gente que estuviera remando (lo encontrareis como Paddling en el DataFrame en la columna activity)?
- 4.3 ¿Cuántos ataques se han producido en Australia antes del 2000?
- 4.4 ¿Y después del 2000?

australia_paddling = australia[australia["activity"] == "Paddling"]
australia_paddling.head()

# Nos devuelve 0 veces

australia_antes_2000 = australia[australia["year"] < 2000]
#australia_antes_2000.info()
australia_antes_2000.head(2)


# 151 casos de ataques

australia_antes_2000 = australia[australia["year"] >= 2000]
#australia_antes_2000.info()
australia_antes_2000.head(2)
# 173 casos

5. Extraed unicamente los ataques a mujeres

ataques_mujeres = fichero[fichero["sex_"]== "F"]
ataques_mujeres.info()
#207 casos mujeres en toda la tabla, lo hemos traido de todos los paises porque hemos interpredado que era en total.

6. ¿Cuántas mujeres fueron atacadas en España?

ataques_mujeres_españa = ataques_mujeres[ataques_mujeres["country"] == "SPAIN"]
ataques_mujeres_españa.info()
#ataques_mujeres_españa

7.Extraed los ataques que se hayan producido en el año 2000, 1997 y 2006. Utilizad el método isin() para resolver este ejercicio.

ataques_x_años = fichero[fichero["year"].isin([2000, 1997, 2006])] 
ataques_x_años.info()

# Intendemos que se piden todos juntos porque .isin pide la lista

8 .En este ejercicio tendréis que extraer los ataques que han ocurrido por tiburón blanco (lo encontraremos en la columna species como algo que contenga white o White). 


Usad el método str.contains() para resolver este ejercicio.

patron = "white|White"

tiburon_blanco = fichero[fichero["species_"].str.contains(patron, regex = True)]
tiburon_blanco.head(2)

9 .Extraed ahora los ataques por tiburón blanco o tigre. Utilizad el método str.contains() para resolver este ejercicio.

patron_2 = "(white|tiger)"

tiburon_blanco_tigre = fichero[fichero["species_"].str.contains(patron_2, case= False, regex = True)]
tiburon_blanco_tigre.head(2)

10. Utilizando la columna de year cread una nueva columna llamada siglo que tiene siglo-XX o siglo-XXI. Donde el año sea menor que 2000 pondremos siglo-XX y el resto de los años siglo-XXI.

fichero["siglo"]= np.where(fichero["year"]>= 2000, "siglo XXI", "siglo XX")
fichero.sample(2)

11. Guardad como un archivo csv nuevo. Lo usaremos en el pair programming de mañana.

fichero.set_index(["case_number"], inplace = True)

fichero.to_csv("DATOS/attacks_pandas4.csv")
