==> PANDAS/5-Unión_de_dataframes.ipynb <==
Trabajaremos con tres ficheros diferentes, para acabar con uno general:
- El csv que creamos en el ejercicio de pair programming Pandas IV.

- El csv llamado attacks_historico donde tenemos información sobre los ataques de tiburón previos al año 1800.
- El csv llamado attacks_adicional donde tenemos información sobre los ataques de tiburón.


import pandas as pd
import numpy as np

1. Cargad los tres ficheros nombrados anteriormente y los exploramos para familizarizarnos con ellos. Escoged una forma de unirlos.

attacks_4 = pd.read_csv("DATOS/attacks_pandas4.csv")

attacks_h = pd.read_csv("DATOS/attacks_historico.csv")

attacks_a = pd.read_csv("DATOS/attacks_adicional.csv")

print(sorted(attacks_4.columns))
print(sorted(attacks_h.columns))
print(sorted(attacks_a.columns))

# Explorando los nombres de las columnas nos damos cuenta de que no están homogeneizados, ya que hay espacios por medio y letras mayúsculas..., y de que la columna 
# "Unnamed: 0" es una columna común en todas, probablemente utilizaremos un merge() para unirlos pero antes queremos unifarmarlo.

2. Chequead que los nombres de las columnas ver si se llaman iguales entre dataframes . ¿Están todas en minúsculas? ¿Tienen espacios? En caso de que no sean todas iguales unificad el nombre de las columnas.

columnas_nuevas_4 = {col : col.lower().replace(" ","_").replace(".","_") for col in attacks_4.columns}
columnas_nuevas_h = {col : col.lower().replace(" ","_").replace(".","_") for col in attacks_h.columns}
columnas_nuevas_a = {col : col.lower().replace(" ","_").replace(".","_") for col in attacks_a.columns}

#print(columnas_nuevas_4)
print(columnas_nuevas_h)
#print(columnas_nuevas_a)

attacks_4.rename(columns= columnas_nuevas_4, inplace = True)
attacks_4.columns

attacks_h.rename(columns= columnas_nuevas_h, inplace = True)
attacks_h.columns

attacks_a.rename(columns= columnas_nuevas_a, inplace = True)
attacks_a.columns

3. En este momento nos habremos dado cuenta que el DataFrame que generamos en el pair programming Pandas IV es similar al DataFrame de los datos historicos. Si recordamos, eliminamos algunas columnas en la sesión III que no eran útiles, pero esas columnas están todavía en el histórico. Si no fuera por eso podríamos unir esos dataframes. Realizad los cambios que creais necesarios para que los dataframes sean iguales y luego juntadlos usando el mejor método.

display(attacks_4.columns, attacks_h.columns)

attacks_h2 = attacks_h.drop(["unnamed:_22", "unnamed:_23", "case_number_1", "case_number_2","original_order", "investigator_or_source","pdf", "href_formula"] ,axis=1)

attacks_h2.columns

4. Es el momento de unir la información de todos los ataques de tiburón que acabamos de crear con el DataFrame de información adicional. Debatid que método es el más correcto para unir toda esta información y ejecutadlo.
- NOTA Cuando hagáis la unión considerad que la información que más nos interesa es la del DataFrame que creamos en el apartado anterior, lo cual determinará el tipo de unión.

#Primero unimos el archivo, el del ejercicio 4 y el attacks histórico.
attacks_4h = attacks_4.merge(attacks_h2, how="outer", on=['activity', 'age', 'area', 'case_number', 'country', 'href', 'location', 'name', 'sex_', 'species_', 'time', 'type', 'unnamed:_0', 'year'])

#Después unimos el resultado anterior con el archivo attacks adicional.
attacks_4ha = attacks_4h.merge(attacks_a, how="outer", on=["case_number", "unnamed:_0"])

attacks_4ha.head(2)

5. Guardad el dataframe nuevo en un csv que usaremos mañana en la clase de pair programming.

attacks_4ha.set_index(["case_number"], inplace=True)

attacks_4ha.to_csv("DATOS/attacks_pandas5.csv")
