==> PANDAS/6-Groupby.ipynb <==
# importamos las librerías que necesitamos

import pandas as pd
import numpy as np

#vamos a trabajar con el csv creado en el ejercicio de pair programming anterior.

fichero = pd.read_csv("DATOS/attacks_pandas5.csv")

1. ¿Cuántos ataques hubo por país? Devuelve un DataFrame.

📌 NOTA Ordena los resultados de mayor a menor. Esto todavía no lo hemos visto, pero un DataFrame lo podemos ordenar con `.sort_values()`.

fichero.head(2)

ataques = fichero.groupby("country").count()
ataques.head()

ataques_ordenados = ataques.sort_values(by = "case_number" ,axis = 0, ascending = False)
ataques_ordenados.head()

2. ¿Cuántos ataques ha habido por año? Devuelve un DataFrame. De nuevo ordena los resultados de mayor a menor. `.sort_values()`



ataques_años = (fichero.groupby("year").count()).sort_values(by = "year" ,axis = 0,  ascending = False)
ataques_años.head()


3. ¿Cuántos ataques hubo por año y sexo? Devolved un DataFrame solo del año 2000 en adelante.

ataques_a_s = fichero.groupby(["year","sex_"])["case_number"].count().reset_index()
ataques_a_s


4. Primero extraed solo las filas que correspondan a USA, AUSTRALIA y SOUTH AFRICA. 

    ¿Cuántos de los ataques fueron letales por sexo? 
    Devuelve un DataFrame.

ataques_uas = fichero[(fichero["country"] == "USA") | (fichero["country"] == "AUSTRALIA") | (fichero["country"] == "SOUTH AFRICA")]
ataques_uas.head(3)

#para comprobar el resultado de los ataques que han sido fatales hemos averiguado la columna fatal_(y/n)

ataques_uas["fatal_(y/n)"].value_counts()


ataques_fatales = ataques_uas[ataques_uas["fatal_(y/n)"]=="Y"].groupby("sex_")["case_number"].count().reset_index()
ataques_fatales

#Hemos entendido que la pregunta es: numero de ataques fatales por sexo.

5. Usando el DataFrame creado en el ejercicio anterior. 
    
    ¿En qué país hubo más ataques?

#En el Dataframe anterior hay solamente los ataques por cada "sexo", por lo cual tendremos que hacer otro aprovechando el codigo del anterior

pais_fatal = ataques_uas[ataques_uas["fatal_(y/n)"]=="Y"].groupby("country")["case_number"].count().reset_index()
pais_fatal.sort_values(by = "case_number" ,axis = 0,  ascending = False) 

6. ¿Qué y cuántos tipos de ataques hubo por país?

tipologia_ataque = fichero.groupby("type")["country"].value_counts()
tipologia_ataque

# Para ver una muestra variada hacemos un sample de los resultados

tipologia_ataque.sample(20)
