==> PANDAS/7-Apply.ipynb <==
import pandas as pd
import numpy as np
import regex as re
import random

fichero = pd.read_csv("DATOS/attacks_pandas5.csv")

1 Cread una columna nueva y una función que nos de el mes en el que ocurrió el ataque. Tened en cuenta que no todas las filas tienen la misma estructura y que puede que no haya la información de mes. En ese caso devolved un nulo (NaN).
💡 Pista 💡Podéis usar regex.

#Averiguamos la tipología de la columna correspondiente a la fecha(date) y vemos que es de tipo object
fichero.info()

# observamos varios ejemplos de como sacan los datos la columna date.
fichero.sample()


def sacar_mes(columna):
    patron = "-(\w+)-"
    #patron2 = "(\d+)-(\w+)-(\d+)"
    try:
        mes =re.findall(patron, columna)
        return mes
          
    except:
        return np.nan
        

fichero["mes"] = fichero["date"].apply(sacar_mes)

fichero.sample(2)

2 Vamos con la columna fatal_(y/n). Lo primero que tenéis que hacer es evaluar los valores únicos que hay. Esperaríamos tener solo dos valores: y y n, pero tristemente no será así 😭. Cread una columna nueva y una función que devuelva únicamente y o n y que devuelva un nulo (NaN) si no se encuentra el valor.

#Sacamos el unique para ver los valores que teníamos en la columna.
fichero["fatal_(y/n)"].unique()

def y_n(columna):
    columna = str(columna)
    columna = columna.upper().replace(" ","")
    if columna == "N":
        return columna
    elif columna == "Y":
        return columna
    else:
        return np.nan


fichero["fatal_(y/n)_limpio"] = fichero["fatal_(y/n)"].apply(y_n)

#Volvemos a sacar el unique para comprobar que la función ha funcionado.
fichero["fatal_(y/n)_limpio"].unique()

3 Seguid la misma lógica que en el ejercicio anterior pero con la columna sex. Pero en este caso vamos a incluir un retito más: En este caso querremos tener dos valores únicos F y M. De nuevo en la columna hay otros carácteres raros como 'M ', 'lli', 'N', '.'. En esos casos, queremos que se reemplacen por M o F de forma aleatoria (tendréis que usar la librería random). En caso de que sea nulo, que la función devuelva nulo (np.nan).

fichero["sex_"].unique()

# Como ya tenemos la columna limpia de resultados hablamos con Cesar y acordamos encontrar la manera de que los nan de esta columna pasen 
# aleatoriamente a "Y" o "N". 

def nulos_aleatorio(colum):
    lista = ["N","Y"]
    if colum == "Y":
        return "Y"
    elif colum == "N":
        return "N"
    else:
        return random.choice(lista)

fichero["sex"] = (fichero["sex_"].apply(nulos_aleatorio))

fichero["sex"].unique()

fichero.sample()

4 Una vez que hayáis terminado eliminad las columnas originales conservando las nuevas.

borrar_columnas = fichero.drop(labels= ["fatal_(y/n)", "sex_"], axis=1)
borrar_columnas

5 Guardad el dataframe en un csv que usaremos más adelante.

fichero.to_csv("DATOS/attacks_pandas7.csv")
